config + clean

+++ Makefile
@@ -1,5 +1,4 @@
 SHELL := /bin/bash
-.DEFAULT_GOAL := help
 .PHONY: *
 .SILENT:
 
@@ -12,55 +11,27 @@ stop:
 	systemctl --user stop kiwix-server.service
 
 status:
+	DBUS_SESSION_BUS_ADDRESS=unix:path=/run/user/$${UID}/bus \
 	systemctl --user status kiwix-server.service
 
-update:
+update_software:
 	podman pull ghcr.io/kiwix/kiwix-serve
 
-download_wikipedia_fr:
-	BASE_URL=https://dumps.wikimedia.org/other/kiwix/zim/wikipedia/; \
-	BASE_FILE=wikipedia_fr_all_nopic; \
-	LATEST=$$(curl --silent $${BASE_URL} | grep $${BASE_FILE} | tail -n 1 | sed -r 's/.*href="([^"]+).*/\1/g'); \
-	if ! [ -f "./zim_files/$${LATEST}" ]; then \
-		wget --directory-prefix=./zim_files/ --output-file=/dev/null $${BASE_URL}$${LATEST}; \
-		rm ./zim_files/$${BASE_FILE}_latest.zim; \
-		ln -s $${LATEST} ./zim_files/$${BASE_FILE}_latest.zim; \
-	fi;
-
-download_wiktionary_fr:
-	BASE_URL=https://dumps.wikimedia.org/other/kiwix/zim/wiktionary/; \
-	BASE_FILE=wiktionary_fr_all_nopic; \
-	LATEST=$$(curl --silent $${BASE_URL} | grep $${BASE_FILE} | tail -n 1 | sed -r 's/.*href="([^"]+).*/\1/g'); \
-	if ! [ -f "./zim_files/$${LATEST}" ]; then \
-		wget --directory-prefix=./zim_files/ --output-file=/dev/null $${BASE_URL}$${LATEST}; \
-		rm ./zim_files/$${BASE_FILE}_latest.zim; \
-		ln -s $${LATEST} ./zim_files/$${BASE_FILE}_latest.zim; \
-	fi;
-
-download_wikiquote_fr:
-	BASE_URL=https://dumps.wikimedia.org/other/kiwix/zim/wikiquote/; \
-	BASE_FILE=wikiquote_fr_all_nopic; \
-	LATEST=$$(curl --silent $${BASE_URL} | grep $${BASE_FILE} | tail -n 1 | sed -r 's/.*href="([^"]+).*/\1/g'); \
-	if ! [ -f "./zim_files/$${LATEST}" ]; then \
-		wget --directory-prefix=./zim_files/ --output-file=/dev/null $${BASE_URL}$${LATEST}; \
-		rm ./zim_files/$${BASE_FILE}_latest.zim; \
-		ln -s $${LATEST} ./zim_files/$${BASE_FILE}_latest.zim; \
-	fi;
-
-download_wikimed_fr:
-	BASE_URL=https://download.kiwix.org/zim/wikipedia/; \
-	BASE_FILE=wikipedia_fr_medicine_maxi; \
-	LATEST=$$(curl --silent $${BASE_URL} | grep $${BASE_FILE} | tail -n 1 | sed -r 's/.*href="([^"]+).*/\1/g'); \
-	if ! [ -f "./zim_files/$${LATEST}" ]; then \
-		wget --directory-prefix=./zim_files/ --output-file=/dev/null $${BASE_URL}$${LATEST}; \
-		rm ./zim_files/$${BASE_FILE}_latest.zim; \
-		ln -s $${LATEST} ./zim_files/$${BASE_FILE}_latest.zim; \
-	fi;
-
-maintenance: update download_wikipedia_fr download_wiktionary_fr download_wikiquote_fr download_wikimed_fr stop start
+update_zim_files:
+	jq --compact-output ".zim_files.[]" config.json | while read SOURCE; do \
+		BASE_URL=$$(jq --raw-output ".base_url" <<< "$${SOURCE}"); \
+		BASE_FILE=$$(jq --raw-output ".base_file" <<< "$${SOURCE}"); \
+		LATEST=$$(curl --silent $${BASE_URL} | grep $${BASE_FILE} | tail -n 1 | sed -r 's/.*href="([^"]+).*/\1/g'); \
+		if ! [ -f "./zim_files/$${LATEST}" ]; then \
+			echo "downloading $${LATEST}"; \
+			wget --directory-prefix=./zim_files/ --output-file=/dev/null $${BASE_URL}$${LATEST}; \
+			[ -f ./zim_files/$${BASE_FILE}_latest.zim ] && rm ./zim_files/$${BASE_FILE}_latest.zim; \
+			ln -s $${LATEST} ./zim_files/$${BASE_FILE}_latest.zim; \
+		fi; \
+	done;
 
 install_prerequisites:
-	sudo apt install -y wget curl podman
+	sudo apt install -y wget curl jq podman
 
 install_service:
 	sudo usermod --add-subuids 100000-165535 --add-subgids 100000-165535 $${USER}
@@ -72,5 +43,8 @@ install_service:
 	systemctl --user daemon-reload
 
 install_cron:
-	echo "0 7 * * * $${USER} cd $$(pwd) && make maintenance 2>&1 | sed \"s|^|\$$(date -Iseconds) |\" >> $$(pwd)/logs/maintenance_cron.log" \
-		| sudo tee /etc/cron.d/kiwix-server_maintenance >/dev/null
+	echo "0 7 * * * $${USER} cd $$(pwd) && make update 2>&1 | sed \"s|^|\$$(date -Iseconds) |\" >> $$(pwd)/logs/updates.cron.log" \
+		| sudo tee /etc/cron.d/kiwix-server_update >/dev/null
+
+update: update_software update_zim_files stop start
+install: install_prerequisites install_service install_cron update
